requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
apify_client==1.8.1
apify_shared==1.1.2
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
environs==11.2.1
fastapi==0.115.5
fastapi_cors==0.0.6
Flask==3.1.0
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.1
more-itertools==10.5.0
packaging==24.2
pydantic==2.10.1
pydantic_core==2.27.1
python-decouple==3.8
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
Werkzeug==3.1.3
